Refactor Admin Dashboard: Extract Token Summary Section into Separate Component
Related to #1128
