+++ src/app/components/admin/index.ts
@@ -1,9 +1,11 @@
 import { Header } from "./Header";
 import { AdminHeaderWithReloadButton } from "./AdminHeaderWithReloadButton";
 import { AdminTabs } from "./AdminTabs";
+import { TokenSummary } from "./TokenSummary";
 
 export {
   Header,
   AdminHeaderWithReloadButton,
   AdminTabs,
+  TokenSummary,
 }
